int --> long for safety

+++ Source/Castro.cpp
@@ -2936,17 +2936,17 @@ Castro::get_numpts ()
      int numpts_1d;
 
      Box bx(geom.Domain());
-     int nx = bx.size()[0];
+     long nx = bx.size()[0];
 
 #if (BL_SPACEDIM == 1)
      numpts_1d = nx;
 #elif (BL_SPACEDIM == 2)
-     int ny = bx.size()[1];
+     long ny = bx.size()[1];
      Real ndiagsq = Real(nx*nx + ny*ny);
      numpts_1d = int(sqrt(ndiagsq))+2*NUM_GROW;  
 #elif (BL_SPACEDIM == 3)
-     int ny = bx.size()[1];
-     int nz = bx.size()[2];
+     long ny = bx.size()[1];
+     long nz = bx.size()[2];
      Real ndiagsq = Real(nx*nx + ny*ny + nz*nz);
      numpts_1d = int(sqrt(ndiagsq))+2*NUM_GROW;
 #endif
